% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigenbenchmark-package.R
\docType{package}
\name{eigenbenchmark-package}
\alias{eigenbenchmark}
\alias{eigenbenchmark-package}
\title{eigenbenchmark: Adapted ATT benchmark for Eigen}
\description{
Adaptation of the ATT benchmark for variable matrix or vector size in Eigen (C++).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pachadotdev/cpp11blaze}
  \item Report bugs at \url{https://github.com/pachadotdev/cpp11blaze/issues}
}

}
\author{
\strong{Maintainer}: Mauricio Vargas Sepulveda \email{m.sepulveda@mail.utoronto.ca} (\href{https://orcid.org/0000-0003-1017-7574}{ORCID})

}
\keyword{internal}
